Moved launchtemplate to providers
